requirements.txt
docopt>=0.6.2
networkx>=1.11
path.py>=8.2
pytest>=2.9.1
pyYAML>=3.11